+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+
+* [#158](https://github.com/babylonlabs-io/finality-provider/pull/158) Remove start height validation
+
 ## v0.12.0
 
 ### Bug Fixes