+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### v0.12.1
+
 ### Bug Fixes
 
 * [#158](https://github.com/babylonlabs-io/finality-provider/pull/158) Remove start height validation